refactor: move registration key to separate file
