+++ sgxvm/src/attestation/tls/helpers.rs
@@ -13,7 +13,7 @@ use crate::attestation::{
     utils::create_attestation_report,
 };
 use crate::attestation::tls::auth::{ClientAuth, ServerAuth};
-use crate::key_manager::{KeyManager, RegistrationKey};
+use crate::key_manager::{KeyManager, keys::RegistrationKey};
 
 /// Prepares config for client side of TLS connection
 pub(super) fn construct_client_config(key_der: Vec<u8>, cert_der: Vec<u8>, is_dcap: bool) -> ClientConfig {
