+++ sgxvm/src/key_manager/mod.rs
@@ -12,7 +12,8 @@ use k256::sha2::{Digest, Sha256 as kSha256};
 
 use crate::error::Error;
 
-pub const REGISTRATION_KEY_SIZE: usize = 32;
+pub mod keys;
+
 pub const SEED_SIZE: usize = 32;
 pub const SEED_FILENAME: &str = ".swtr_seed";
 pub const PUBLIC_KEY_SIZE: usize = 32;
@@ -341,7 +342,7 @@ impl KeyManager {
     /// Encrypts master key using shared key
     pub fn to_encrypted_master_key(
         &self,
-        reg_key: &RegistrationKey,
+        reg_key: &keys::RegistrationKey,
         public_key: Vec<u8>,
     ) -> Result<Vec<u8>, Error> {
         // Convert public key to appropriate format
@@ -365,7 +366,7 @@ impl KeyManager {
 
     /// Recovers encrypted master key obtained from seed exchange server
     pub fn from_encrypted_master_key(
-        reg_key: &RegistrationKey,
+        reg_key: &keys::RegistrationKey,
         public_key: Vec<u8>,
         encrypted_master_key: Vec<u8>,
     ) -> Result<Self, Error> {
@@ -428,48 +429,6 @@ impl KeyManager {
     //     let public_key = x25519_dalek::PublicKey::from(public_key);
     //     Ok(public_key)
     // }
-}
-
-/// RegistrationKey handles all operations with registration key such as derivation of public key,
-/// derivation of encryption key, etc.
-pub struct RegistrationKey {
-    inner: [u8; REGISTRATION_KEY_SIZE],
-}
-
-impl RegistrationKey {
-    /// Generates public key for seed sharing
-    pub fn public_key(&self) -> x25519_dalek::PublicKey {
-        let secret = x25519_dalek::StaticSecret::from(self.inner);
-        x25519_dalek::PublicKey::from(&secret)
-    }
-
-    /// Generates random registration key
-    pub fn random() -> SgxResult<Self> {
-        // Generate random seed
-        let mut buffer = [0u8; REGISTRATION_KEY_SIZE];
-        let res = unsafe { sgx_read_rand(&mut buffer as *mut u8, REGISTRATION_KEY_SIZE) };
-
-        match res {
-            sgx_status_t::SGX_SUCCESS => Ok(Self { inner: buffer }),
-            _ => {
-                println!(
-                    "[KeyManager] Cannot generate random registration key. Reason: {:?}",
-                    res.as_str()
-                );
-                Err(res)
-            }
-        }
-    }
-
-    /// Performes Diffie-Hellman derivation of encryption key for master key encryption
-    /// * public_key - User public key
-    pub fn diffie_hellman(
-        &self,
-        public_key: x25519_dalek::PublicKey,
-    ) -> x25519_dalek::SharedSecret {
-        let secret = x25519_dalek::StaticSecret::from(self.inner);
-        secret.diffie_hellman(&public_key)
-    }
 }
 
 /// Tries to return path to $HOME/.swisstronik-enclave directory.
